+++ include/filerepeater/VectorToTxtFile.h
@@ -703,7 +703,7 @@ namespace gr {
        * class. filerepeater::VectorToTxtFile::make is the public interface for
        * creating new instances.
        */
-      static sptr make(const char *filename, int vectorLen, float frequency, float sampleRate, const char *notes, bool append, float updateRateSec, int precision, bool WriteTimeHeader);
+      static sptr make(const char *filename, int vectorLen, double frequency, double sampleRate, const char *notes, bool append, float updateRateSec, int precision, bool WriteTimeHeader);
     };
 
   } // namespace filerepeater
